Fix bug 3420884 "Support setMethods with an assignable parameter type".
The functionality of MethodCache.getSetMethod is left unchanged as much as possible. Only if there's no matching method and also no registered autoboxing converter all methods of the objects class are checked if they declare an assignable parameter type. If there's no such method but some with a matching name, they are reported to the user (with the full method signature).